Organize builder into webmock namespace
