+++ lib/stub_requests.rb
@@ -14,22 +14,26 @@ require "stub_requests/version"
 require "stub_requests/core_ext"
 require "stub_requests/hash_util"
 require "stub_requests/argument_validation"
+
 require "stub_requests/endpoint"
 require "stub_requests/endpoint_registry"
+
 require "stub_requests/metrics"
 require "stub_requests/metrics/endpoint_stat"
 require "stub_requests/metrics/registry"
 require "stub_requests/metrics/stub_stat"
+
 require "stub_requests/service"
 require "stub_requests/service_registry"
+
 require "stub_requests/uri"
 require "stub_requests/uri/scheme"
 require "stub_requests/uri/suffix"
 require "stub_requests/uri/validator"
 require "stub_requests/uri/builder"
+
+require "stub_requests/webmock/builder"
 require "stub_requests/webmock/stub_registry_extension"
-require "stub_requests/webmock_builder"
 
 require "stub_requests/api"
-
 require "stub_requests/stub_requests"
